Moving code around as I see fit
